Refactor LZO compression stuff into its own util file
